requirements.txt
telethon
python-dotenv
psutil
opencv-python-headless
devgagantools
aiofiles
https://www.dl.dropboxusercontent.com/scl/fi/e0fo6fcjn8kmr5r0x6wvg/myownpyro.zip?rlkey=d1znpwckss4ullz0sg7e1qjjg&st=kmbh7wdv&dl=0
telethon-tgcrypto
motor
pymongo
pytz
Pillow
tgcrypto
cryptography
